requirements.txt
# GraphRAG系统依赖包

# Web框架
Flask==2.3.3
Werkzeug==2.3.7

# 数据库
PyMySQL==1.1.0
SQLAlchemy==2.0.23
neo4j==5.14.1
pymilvus==2.3.4

# 配置和工具
PyYAML==6.0.1
numpy==1.24.3

# 机器学习
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2

# 文档处理
unstructured==0.11.0
unstructured[pdf]==0.11.0

# 其他工具
requests==2.31.0
python-dotenv==1.0.0 